Common code step 1
